Versao producao - incluido funcionalidade de inclusao de proposta na tela inicial. Incluido novo layout

+++ DWM-Imovel/DWM-Imovel/Models/Repositories/PropostaViewModel.cs
@@ -78,9 +78,9 @@ namespace DWM.Models.Repositories
         [DisplayName("Login")]
         public string login { get; set; }
 
-        public decimal percent_atual { get; set; }
+        public double percent_atual { get; set; }
 
-        public decimal percent_restnte { get; set; }
+        public double percent_restnte { get; set; }
     }
 }
